--- backend/baguni-api/Dockerfile_Selenium
@@ -1,18 +1,0 @@
-FROM openjdk:17-jdk-slim
-
-# 1. 필수 패키지 업데이트 및 설치
-RUN apt-get update -y && \
-    apt-get install -y wget unzip
-
-# 2. 크롬 설치
-RUN wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb && \
-    apt-get -y install ./google-chrome-stable_current_amd64.deb && \
-    rm -f ./google-chrome-stable_current_amd64.deb && \
-    apt-get clean && \
-    rm -rf /var/lib/apt/lists/*
-
-# 3. ChromeDriver 설치
-ARG CHROME_VERSION=114
-RUN wget -O /tmp/chromedriver.zip https://chromedriver.storage.googleapis.com/$(curl -sS chromedriver.storage.googleapis.com/LATEST_RELEASE_${CHROME_VERSION})/chromedriver_linux64.zip && \
-    unzip /tmp/chromedriver.zip chromedriver -d /usr/bin && \
-    rm -f /tmp/chromedriver.zip
