+++ backend/baguni-api/src/main/java/baguni/api/application/link/dto/LinkApiMapper.java
@@ -4,7 +4,7 @@ import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 
 @Mapper(
 	componentModel = "spring",
