+++ backend/baguni-api/src/main/java/baguni/api/application/pick/dto/PickApiRequest.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 
 public class PickApiRequest {
