+++ backend/baguni-api/src/main/java/baguni/api/application/pick/dto/PickApiResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import jakarta.validation.constraints.NotNull;
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 import baguni.api.service.pick.dto.PickResult;
 
 public class PickApiResponse {
