+++ backend/baguni-api/src/main/java/baguni/api/infrastructure/pick/PickBulkDataHandler.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 import baguni.api.service.pick.dto.PickCommand;
 import baguni.entity.model.link.Link;
 import baguni.entity.model.link.LinkRepository;
