+++ backend/baguni-api/src/main/java/baguni/api/infrastructure/pick/PickDataHandler.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import baguni.api.service.folder.exception.ApiFolderException;
-import baguni.api.service.link.dto.LinkMapper;
+import baguni.entity.model.link.dto.LinkMapper;
 import baguni.api.service.pick.dto.PickCommand;
 import baguni.api.service.pick.dto.PickMapper;
 import baguni.api.service.pick.exception.ApiPickException;
