+++ backend/baguni-api/src/main/java/baguni/api/service/pick/dto/PickCommand.java
@@ -2,7 +2,7 @@ package baguni.api.service.pick.dto;
 
 import java.util.List;
 
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 
 public class PickCommand {
