+++ backend/baguni-api/src/test/java/baguni/api/domain/pick/service/PickServiceTest.java
@@ -24,7 +24,7 @@ import baguni.entity.model.user.SocialProvider;
 import lombok.extern.slf4j.Slf4j;
 import baguni.BaguniApiApplication;
 import baguni.api.application.pick.dto.PickApiMapper;
-import baguni.api.service.link.dto.LinkInfo;
+import baguni.entity.model.link.dto.LinkInfo;
 import baguni.api.service.pick.dto.PickCommand;
 import baguni.api.service.pick.dto.PickResult;
 import baguni.api.service.pick.exception.ApiPickException;
