+++ backend/baguni-entity/src/main/java/baguni/entity/model/link/dto/LinkResult.java
@@ -1,4 +1,4 @@
-package baguni.api.service.link.dto;
+package baguni.entity.model.link.dto;
 
 import java.time.LocalDateTime;
